src/web/requirements/base.txt
# latest django
Django==1.6.1

# models
South==0.8.4
django-model-utils==1.5.0

# date time utils
pytz==2013.8
python-dateutil==2.2

# rest api

# misc
six==1.4.1
wsgiref==0.1.2

# rest api
djangorestframework==2.3.10
Markdown==2.3.1
django-filter==0.7

